server/requirements.txt
fastapi[standard]
rembg[cpu]
python-multipart
pillow